+++ dd-java-agent/instrumentation/mule-4/src/main/java/datadog/trace/instrumentation/mule4/ExecutionInitialSpanInfoInstrumentation.java
@@ -23,6 +23,11 @@ public class ExecutionInitialSpanInfoInstrumentation extends InstrumenterModule.
   }
 
   @Override
+  protected boolean defaultEnabled() {
+    return false;
+  }
+
+  @Override
   public String instrumentedType() {
     return "org.mule.runtime.tracer.customization.impl.info.ExecutionInitialSpanInfo";
   }
